fix(move func): arrangement of functions
**What**

- logically correct arrangement of functions

closed #167
