+++ backend/server/blueprints/admin_calendars_id_delete.py
