+++ backend/server/endpoints_func/add_slot_from_db_for_schedule_admin.py
