+++ backend/server/endpoints_func/delete_slot_by_id_for_schedule.py
