+++ backend/server/endpoints_func/get_booking_settings.py
