+++ backend/server/endpoints_func/get_slots_by_filter.py
