+++ backend/server/endpoints_func/query_slots.py
