+++ backend/tests/test_calendars_post_for_admin.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta
-from server.calendars_post_for_admin import get_expiry_date
+from server.blueprints.calendars_post_for_admin import get_expiry_date
 from flask import json
-from tools.datetime_convertations import DateTime
+from tools.datetime.datetime_convertations import DateTime
 
 
 def test_expiry_date_30():
