+++ backend/tests/test_error_handler.py
@@ -1,5 +1,5 @@
 import pytest
-from tools.create_db_for_tests import create_test_app_with_db
+from tools.for_test.create_db_for_tests import create_test_app_with_db
 from tools.for_db.work_with_admin_info import add_admin
 from tools.func_for_psw import password_hashing
 import base64
