+++ backend/tests/test_post_schedule_endpoint.py
@@ -1,5 +1,5 @@
 import base64
-from tools.create_db_for_tests import create_test_app_with_db
+from tools.for_test.create_db_for_tests import create_test_app_with_db
 from tools.for_db.work_with_admin_info import add_admin
 from tools.func_for_psw import password_hashing
 import pytest
