+++ backend/tools/for_db/work_with_links.py
@@ -1,5 +1,5 @@
 from db import models
-from tools.datetime_convertations import DateTime
+from tools.datetime.datetime_convertations import DateTime
 from db.models import Links, LinksSchema, db
